# openapi.yaml
openapi: 3.1.0
info:
  title: FindTutor API
  version: 0.1.0
  description: >
    Contract for the FindTutor language-tutoring marketplace (Uzbekistan).
    Hand-written and treated as the source of truth: the Go backend renders
    responses to match it, and the Next.js frontend generates its typed client
    from it (openapi-typescript + openapi-fetch).

    Money is always an integer amount in MINOR units (tiyin for UZS, cents for
    USD) plus a currency code — never a float.

    Language: every human-readable `error.message` is rendered in the language
    of the request's `Accept-Language` header (`en`, `ru` or `uz`; anything
    else falls back to English) and the response carries `Content-Language`.
    Email is written in the account's stored `locale` (see AuthUser), set from
    Accept-Language at registration and changed via PATCH /v1/auth/me.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/bookings:
    post:
      tags: [bookings]
      summary: Book a lesson
      description: >
        Creates a booking in `pending_payment` and opens a `requires_payment`
        payment intent for its price. The server re-derives the bookable slot
        set and rejects a `start_at` that is not a real, currently free slot; it
        never trusts a client-supplied price.

        `duration_minutes` must be one of 30/60/90/120. When `is_trial` is true
        the duration is forced to 30 minutes and the price is the teacher's
        `trial_price_minor` (400 if the teacher offers no trial). You cannot
        book your own teacher profile.

        The student then calls `POST /v1/bookings/{id}/pay` to authorise
        payment, which confirms the booking.
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingRequest" }
      responses:
        "201":
          description: The created booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >
            The slot is not bookable (`slot_unavailable` — outside availability,
            misaligned, in the past, or already taken) or was taken by a
            concurrent request between the check and the insert (`slot_taken`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [bookings]
      summary: List the caller's bookings
      description: >
        With no `role`, returns every booking the caller participates in (as
        student or as the owner of the teacher profile). `role=student` limits
        it to bookings the caller booked; `role=teacher` to bookings for the
        teacher profile the caller owns (empty if they own none). Newest lesson
        first. Optional `status` filter.
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema: { type: string, enum: [student, teacher] }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/BookingStatus" }
      responses:
        "200":
          description: The matching bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingList" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/bookings/{id}/disputes:
    post:
      tags: [disputes]
      summary: Raise a dispute on a lesson
      description: >
        Participant-only — the student or the teacher-owner on the booking (403
        otherwise). The lesson must be `confirmed` or `completed` (409
        `dispute_not_allowed`), and a booking may have only one OPEN dispute at a
        time (409 `dispute_exists`, enforced by a partial unique index, so a
        concurrent double-submit is race-safe). Once a dispute is resolved or
        rejected a new one may be raised; the whole thread is kept.

        `reason` is required, trimmed, and capped at 2000 characters.
      operationId: raiseBookingDispute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RaiseDisputeRequest" }
      responses:
        "201":
          description: The created dispute.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispute" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not a participant in this booking (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            The lesson is not `confirmed` / `completed` (`dispute_not_allowed`),
            or it already has an open dispute (`dispute_exists`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [disputes]
      summary: A booking's dispute thread
      description: >
        Participant-only. Every dispute ever raised on the booking, open and
        closed, newest first.
      operationId: listBookingDisputes
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The dispute thread (possibly empty).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DisputeList" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The caller is not a participant in this booking (`forbidden`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/admin/bookings:
    get:
      tags: [admin]
      summary: Every booking on the platform
      description: >
        Permission: `bookings.view`. Unlike `GET /v1/bookings` (the caller's own
        lessons) this lists every booking, newest lesson first. `status` is an
        optional exact filter; `q` matches the teacher's display name or slug and
        the student's email or display name, case-insensitively.
      operationId: adminListBookings
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/BookingStatus" } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminBookingList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /v1/admin/disputes:
    get:
      tags: [admin]
      summary: The dispute queue
      description: >
        Permission: `disputes.resolve`. Disputes newest first, each with the
        booking and both parties. `status` defaults to `open`; pass `all` to
        include closed disputes.
      operationId: adminListDisputes
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [open, resolved, rejected, all], default: open } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of disputes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDisputeList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/disputes/{id}/resolve:
    post:
      tags: [admin]
      summary: Resolve or reject a dispute
      description: >
        Permission: `disputes.resolve`. Closes an OPEN dispute, recording the
        outcome, the required `resolution` note (trimmed, at most 2000
        characters), the operator and the timestamp. 409 `already_resolved` when
        the dispute is no longer open — the UPDATE is guarded on the open status,
        so a lost race is reported rather than overwriting the other operator's
        resolution.

        With `outcome: resolved` and `refund: true` the booking's payment is
        refunded; the booking's own lifecycle is deliberately left alone (use
        force-cancel to also cancel the lesson). `refund` is ignored for
        `outcome: rejected`.
      operationId: adminResolveDispute
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveDisputeRequest" }
      responses:
        "200":
          description: The closed dispute.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dispute" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: "`dispute_not_found`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: "`already_resolved` — the dispute is no longer open."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived HS256 access token from /v1/auth/{login,register,refresh}.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ftr_session
      description: >
        HttpOnly opaque refresh token, Path=/v1/auth, SameSite=Lax. Set by the
        auth endpoints; the browser sends it automatically to /v1/auth/refresh
        and /v1/auth/logout. Not readable or settable by client JS.
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The caller is not allowed to perform this action.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable identifier (e.g. `not_found`).
            message:
              type: string
              description: Human-readable explanation, safe to surface to users.
...
    Currency:
      type: string
      enum: [UZS, USD]
    Money:
      type: object
      required: [amount_minor, currency]
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Amount in minor units (tiyin / cents). 9000000 UZS == 90,000 so'm.
        currency: { $ref: "#/components/schemas/Currency" }
...
    BookingStatus:
      type: string
      enum: [pending_payment, confirmed, completed, cancelled]
...
    BookingTeacherSummary:
      type: object
      description: Light teacher summary embedded in a booking so lists avoid N+1 calls.
      required: [slug, display_name, timezone, avatar_url]
      properties:
        slug: { type: string }
        display_name: { type: string }
        timezone: { type: string, description: IANA name. }
        avatar_url: { type: string }
    BookingStudentSummary:
      type: object
      required: [id, display_name]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
    Booking:
      type: object
      required:
        - id
        - status
        - start_at
        - end_at
        - duration_minutes
        - is_trial
        - price
        - created_at
        - cancelled_at
        - cancelled_by
        - no_show_party
        - can_review
        - review
        - can_raise_dispute
        - open_dispute
        - teacher
        - student
        - payment
        - resources
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BookingStatus" }
        start_at: { type: string, format: date-time, description: RFC3339 UTC. }
        end_at: { type: string, format: date-time, description: RFC3339 UTC. }
        duration_minutes: { type: integer }
        is_trial: { type: boolean }
        price: { $ref: "#/components/schemas/Money" }
        created_at: { type: string, format: date-time }
        cancelled_at:
          type: string
          format: date-time
          nullable: true
          description: Set when status is `cancelled`, otherwise null.
        cancellation_reason:
          type: string
          description: Present (may be empty) only once cancelled; omitted otherwise.
        cancelled_by:
          type: string
          enum: ["", student, teacher, admin]
          description: >
            Empty until the booking is cancelled, then who cancelled it: the
            `student`, the `teacher` (a teacher no-show counts as a teacher
            cancellation), or an `admin` force-cancel.
        meeting_url:
          type: string
          description: >
            The effective video link (the per-booking override if set, else the
            teacher's default `meeting_url`). Present ONLY when the caller is a
            participant AND `status` is `confirmed` or `completed`; omitted for
            everyone else (it never leaks to a `pending_payment` booking or a
            non-participant).
        no_show_party:
          type: string
          enum: ["", student, teacher]
          description: >
            Empty normally. `student` when the student missed a lesson the
            teacher attended (booking `completed`, payment captured); `teacher`
            when the teacher self-reported they could not attend (booking
            `cancelled`, student refunded).
        can_review:
          type: boolean
          description: >
            True only when the caller is the student, `status` is `completed`,
            and no review exists yet — the frontend shows the post-lesson review
            prompt off this without a second call.
        review:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/BookingReview" }]
          description: >
            The caller's review of this booking if one exists (visible to both
            participants); null otherwise.
        can_raise_dispute:
          type: boolean
          description: >
            True only when the caller is a participant, `status` is `confirmed`
            or `completed`, and no dispute is currently open — the frontend shows
            the "report a problem" action off this without a second call.
        open_dispute:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/BookingOpenDispute" }]
          description: >
            The booking's currently open dispute (visible to both participants);
            null when there is none. The full thread, including closed disputes,
            is at `GET /v1/bookings/{id}/disputes`.
        teacher: { $ref: "#/components/schemas/BookingTeacherSummary" }
        student: { $ref: "#/components/schemas/BookingStudentSummary" }
        payment:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/BookingPayment" }]
          description: >
            The booking's payment intent. Null when there is none yet, and
            omitted from list responses (only single-booking responses embed
            it).
        resources:
          type: array
          items: { $ref: "#/components/schemas/BookingResource" }
          description: >
            The lesson's attached materials / homework, summary only (no quiz
            content — see GET /v1/bookings/{id}/resources for that). Always an
            array, never null.
    BookingPayment:
      type: object
      description: The payment intent embedded in a booking.
      required: [status, amount_minor, currency]
      properties:
        status:
          type: string
          enum: [requires_payment, authorized, captured, refunded, failed]
        amount_minor: { type: integer, format: int64, description: Integer minor units. }
        currency: { type: string }
...
    BookingList:
      type: object
      required: [bookings]
      properties:
        bookings:
          type: array
          items: { $ref: "#/components/schemas/Booking" }
    CreateBookingRequest:
      type: object
      required: [teacher_slug, start_at, duration_minutes]
      properties:
        teacher_slug: { type: string }
        start_at:
          type: string
          format: date-time
          description: RFC3339 UTC. Must be in the future and a real bookable slot.
        duration_minutes:
          type: integer
          enum: [30, 60, 90, 120]
          description: Ignored when is_trial is true (forced to 30).
        is_trial:
          type: boolean
          default: false
          description: A trial lesson — 30 minutes, priced at the teacher's trial_price.
...
    BookingReview:
      type: object
      description: The review attached to a booking, embedded in booking responses.
      required: [rating, comment, created_at]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        created_at: { type: string, format: date-time, description: RFC3339 UTC. }
...
    ResourceType:
      type: string
      enum: [material, article, quiz, listening, reading, writing]
...
    BookingResource:
      type: object
      description: >
        The summary of one resource attached to a booking, embedded in
        `Booking.resources`. No quiz content — see GET
        /v1/bookings/{id}/resources for the full attach view.
      required: [id, resource_id, kind, position, due_at, type, title, resource_status, submission_status]
      properties:
        id: { type: string, format: uuid, description: The attachment id. }
        resource_id: { type: string, format: uuid }
        kind: { type: string, enum: [material, homework] }
        position: { type: integer }
        due_at: { type: string, format: date-time, nullable: true }
        type: { $ref: "#/components/schemas/ResourceType" }
        title: { type: string }
        resource_status: { type: string, enum: [draft, published] }
        submission_status:
          type: string
          enum: ["", in_progress, submitted, graded]
          description: >
            The viewer's own submission status for this homework; empty when
            the viewer is not the student, or hasn't started it yet.
...
    DisputeStatus:
      type: string
      enum: [open, resolved, rejected]
      description: "`open` is the only non-terminal state."
    DisputeUser:
      type: object
      description: The account that raised or resolved a dispute.
      required: [id, display_name]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
    BookingOpenDispute:
      type: object
      description: The open dispute embedded in a participant's booking response.
      required: [id, status, reason, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/DisputeStatus" }
        reason: { type: string }
        created_at: { type: string, format: date-time, description: RFC3339 UTC. }
    Dispute:
      type: object
      required: [id, booking_id, status, reason, resolution, raised_by, resolved_by, created_at, resolved_at]
      properties:
        id: { type: string, format: uuid }
        booking_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/DisputeStatus" }
        reason:
          type: string
          description: The participant's account of the problem, at most 2000 characters.
        resolution:
          type: string
          description: The operator's closing note. Empty while the dispute is open.
        raised_by: { $ref: "#/components/schemas/DisputeUser" }
        resolved_by:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/DisputeUser" }]
          description: The operator who closed it; null while open.
        created_at: { type: string, format: date-time, description: RFC3339 UTC. }
        resolved_at:
          type: string
          format: date-time
          nullable: true
          description: RFC3339 UTC; null while open.
    DisputeList:
      type: object
      required: [disputes]
      properties:
        disputes:
          type: array
          items: { $ref: "#/components/schemas/Dispute" }
          description: Newest first.
    RaiseDisputeRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 2000
          description: Trimmed; must be non-empty.
    ResolveDisputeRequest:
      type: object
      required: [outcome, resolution]
      properties:
        outcome:
          type: string
          enum: [resolved, rejected]
          description: >
            `resolved` sides with the person who raised it, `rejected` closes it
            without action.
        resolution:
          type: string
          minLength: 1
          maxLength: 2000
          description: The operator's closing note. Trimmed; must be non-empty.
        refund:
          type: boolean
          default: false
          description: >
            Refund the booking's payment. Only honoured with
            `outcome: resolved`; the booking's own status is not changed.

    # --- Admin (phase A/B) ---
...
    AdminBookingTeacher:
      type: object
      required: [slug, display_name]
      properties:
        slug: { type: string }
        display_name: { type: string }
    AdminBookingStudent:
      type: object
      required: [id, email, display_name]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
    AdminBookingRow:
      type: object
      required: [id, status, start_at, end_at, teacher, student, price, payment_status, created_at, has_open_dispute]
      properties:
        id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/BookingStatus" }
        start_at: { type: string, format: date-time, description: RFC3339 UTC. }
        end_at: { type: string, format: date-time, description: RFC3339 UTC. }
        teacher: { $ref: "#/components/schemas/AdminBookingTeacher" }
        student: { $ref: "#/components/schemas/AdminBookingStudent" }
        price: { $ref: "#/components/schemas/Money" }
        payment_status:
          type: string
          nullable: true
          enum: [requires_payment, authorized, captured, refunded, failed, null]
          description: Null when the booking has no payment intent yet.
        created_at: { type: string, format: date-time }
        has_open_dispute:
          type: boolean
          description: True while the booking has a dispute in the `open` state.
    AdminBookingList:
      type: object
      required: [bookings, total]
      properties:
        bookings: { type: array, items: { $ref: "#/components/schemas/AdminBookingRow" } }
        total: { type: integer, description: Total matches, not just this page. }
...
    AdminDisputeRow:
      description: A queued dispute plus the booking and parties it is about.
      allOf:
        - $ref: "#/components/schemas/Dispute"
        - type: object
          required: [booking]
          properties:
            booking:
              type: object
              required: [id, status, start_at, price, teacher, student]
              properties:
                id: { type: string, format: uuid }
                status: { $ref: "#/components/schemas/BookingStatus" }
                start_at: { type: string, format: date-time }
                price: { $ref: "#/components/schemas/Money" }
                teacher: { $ref: "#/components/schemas/AdminBookingTeacher" }
                student: { $ref: "#/components/schemas/AdminBookingStudent" }
    AdminDisputeList:
      type: object
      required: [disputes, total]
      properties:
        disputes: { type: array, items: { $ref: "#/components/schemas/AdminDisputeRow" } }
        total: { type: integer, description: Total matches, not just this page. }

    # --- Admin payouts (phase E) ---
...